requirements.txt
python-dotenv
pypdf
faiss-cpu
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain_community